Refactor the code structure with a drink folder inside manage, also generalized firebase interactions to allow flexibility in the future.
